improve format and details
